moved config into core/config
